+++ template/public/auth.js
@@ -4,36 +4,33 @@
 
 const auth = {
 
-  loggedIn() {
-    const user = JSON.parse(localStorage.getItem('user'))
+  USER_KEY: '__user__',
+
+  isLoginIn() {
+    const user = JSON.parse(localStorage.getItem(auth.USER_KEY))
     // 用户处于登录状态的条件：本地存储以及 cookie 同时存在
-    if (user && auth.isUserInCookie()) {
-      auth.register(user)
+    if (user) {
+      auth.user = user
       return true
     } else {
       return false
     }
   },
 
-  isUserInCookie() {
-    return document.cookie.match(/(^|; )user=/)
-  },
-
   register(user) {
     auth.user = user
     // 用户信息（基本信息、权限等存放 localStorage，减少前后端通信）
-    localStorage.setItem('user', JSON.stringify(user))
+    localStorage.setItem(auth.USER_KEY, JSON.stringify(user))
   },
 
   destroy() {
     auth.user = null
-    localStorage.removeItem('user')
+    localStorage.removeItem(auth.USER_KEY)
   }
 }
 
 if (__DEV__) {
   // 开发环境登录状态写死，方便开发
-  document.cookie = 'user=0FAC6FBASDS3213AX'
   auth.register({
     name: 'DEMO'
   })