+++ template/webpack.config.js
@@ -42,9 +42,6 @@ var config = {
         plugins: ['transform-runtime']
       }
     }, {
-      test: require.resolve("pace-progress"),
-      loader: "imports?define=>false"
-    }, {
       test: /\.css$/,
       loader: 'style!css'
     }, {